build fix

+++ wallet/wallet_db.h
@@ -51,7 +51,7 @@ namespace beam
 		virtual int getVarRaw(const char* name, void* data) const = 0;
         virtual Height getCurrentHeight() const = 0;
 
-        virtual std::vector<TxDescription> getTxHistory(uint64_t start = 0, int count = std::numeric_limits<size_t>::max()) = 0;
+        virtual std::vector<TxDescription> getTxHistory(uint64_t start = 0, int count = std::numeric_limits<int>::max()) = 0;
         virtual boost::optional<TxDescription> getTx(const Uuid& txId) = 0;
         virtual void saveTx(const TxDescription& p) = 0;
         virtual void deleteTx(const Uuid& txId) = 0;
